requirements.txt
torch
torchvision
numpy
pillow
streamlit
opencv-python
efficientnet_pytorch
